[musi][streama] Rollback & pin h2 version
I think it was not working since NixOS 22.05…
It took time to understand that something like
```nix
pkgs.h2 = super.h2.overrideAttrs { version = OLDVERSION }
```
Didn't really override the version, maybe because of a recursive set in
the derivation.

So I just took copy/pasted the last version of the h2 derivation code.