[musi/picasoft] a temporary pedagogical container
This permits participants of the course to have access to an online
server and to fiddle with unix permissions together, give one to another
access to a directory, have a directory accessible on the web and get a
grasp of the self-host world

most of nginx conf is in the inner-nginx. Outer nginx just proxies the
inner one.

an index.html page is written, listing all the user directories.

a directory is made, owned and writable by the group all students are
in.